chore: moved filter one component higher
